added editing products

+++ src/main/java/ru/sixzr/repositories/ProductRepository.java
@@ -3,5 +3,5 @@ package ru.sixzr.repositories;
 import org.springframework.data.repository.CrudRepository;
 import ru.sixzr.models.Product;
 
-public interface ProductRepository extends CrudRepository<Product, Integer> {
+public interface ProductRepository extends CrudRepository<Product, Long> {
 }
